Code changes
Changed Author in line 5 to correctly represent myself on Github.
Swapped from comments to docstrings for readability throughout entire program to date.

moved some code to EOF for readability.

README updated to explain thoroughly what this program is and my expectations for it, as well as how to use it. Also explains the program is a work in progress. Name may change in the future.