Created branch feature, updated code
Still learning GitHub, was told I should make this branch for code that is still being developed. So here we are. In the future all changes will be done to here first and then reviewed.

Line 99/105: renamed vague sql_table to accuractely reflect what it is. Now -monthly_table-
Line 101: replaced .format with an f-string for readability.